+++ packages/gatsby-telemetry/src/telemetry.ts
@@ -14,6 +14,7 @@ import { createFlush } from "./create-flush"
 import { EventStorage } from "./event-storage"
 import { showAnalyticsNotification } from "./show-analytics-notification"
 import { cleanPaths } from "./error-helpers"
+import { getDependencies } from "./get-dependencies"
 
 import { join, sep } from "path"
 import isDocker from "is-docker"
@@ -53,6 +54,7 @@ export interface IAggregateStats {
 interface IAnalyticsTrackerConstructorParameters {
   componentId?: SemVer
   gatsbyCliVersion?: SemVer
+  trackingEnabled?: boolean
 }
 
 export interface IStructuredError {
@@ -92,6 +94,8 @@ export interface ITelemetryTagsPayload {
   error?: IStructuredError | Array<IStructuredError>
   cacheStatus?: string
   pluginCachePurged?: string
+  dependencies?: Array<string>
+  devDependencies?: Array<string>
   siteMeasurements?: {
     pagesCount?: number
     clientsCount?: number
@@ -132,11 +136,15 @@ export class AnalyticsTracker {
   constructor({
     componentId,
     gatsbyCliVersion,
+    trackingEnabled,
   }: IAnalyticsTrackerConstructorParameters = {}) {
     this.componentId = componentId || `gatsby-cli`
     try {
       if (this.store.isTrackingDisabled()) {
         this.trackingEnabled = false
+      }
+      if (trackingEnabled !== undefined) {
+        this.trackingEnabled = trackingEnabled
       }
 
       this.defaultTags = this.getTagsFromEnv()
@@ -149,6 +157,7 @@ export class AnalyticsTracker {
       // ignore
     }
     this.machineId = this.getMachineId()
+    this.captureMetadataEvent()
   }
 
   // We might have two instances of this lib loaded, one from globally installed gatsby-cli and one from local gatsby.
@@ -461,6 +470,19 @@ export class AnalyticsTracker {
       stdDev: stdDev,
       skewness: !Number.isNaN(skewness) ? skewness : 0,
     }
+  }
+
+  captureMetadataEvent(): void {
+    if (!this.isTrackingEnabled()) {
+      return
+    }
+    const deps = getDependencies()
+    const evt = {
+      dependencies: deps?.dependencies,
+      devDependencies: deps?.devDependencies,
+    }
+
+    this.captureEvent(`METADATA`, evt)
   }
 
   async sendEvents(): Promise<boolean> {